# Codechange: Use vehicle tile hash to build depot vehicle list.