<!--
Commit message:

- Please use Feature / Add / Change / Fix for player-facing changes. E.g.: "Feature: My cool new feature".
- Please use Feature / Add / Change / Fix followed by "[NewGRF]" or "[Script]" for moddable changes. E.g.: "Feature: [NewGRF] My cool new NewGRF addition".
- Please use Codechange / Codefix for developer-facing changes. E.g.: "Codefix #1234: Validate against nullptr properly".

See https://github.com/OpenTTD/OpenTTD/blob/master/CODINGSTYLE.md#commit-message for more details.
-->

## Motivation / Problem

`BuildDepotVehicleList()` iterates the vehicle pool, and the first things it checks are vehicle type and tile. This means it can easily be improved (performance-wise) by using the vehicle tile hash instead.

Depot vehicle lists are used when performing bulk depot commands, as well as for displaying the vehicles in a depot.

<!--
Describe here shortly
* For bug fixes:
    * What problem does this solve?
    * If there is already an issue, link the issue, otherwise describe the problem here.
* For features or gameplay changes:
    * What was the motivation to develop this feature?
    * Does this address any problem with the gameplay or interface?
    * Which group of players do you think would enjoy this feature?
-->


## Description

Use `FindVehicleOnPos()` to build the depot vehicle list instead of iterating the entire vehicle pool.

Observed performance measurements:

| Save           | Depot | Master   | This PR |
| -------------- | ----- | -------- | ------- |
| Wentbourne     | Train | 1,073us  | 7us+    |
| Wentbourne     | Ship  | 714us    | 7us+    |
| Xarick's 4kx4k | Train | 13,386us | 25us+   |
| Xarick's 4kx4k | Ship  | 8,734us  | 25us+   |
| Xarick's 4kx4k | Road  | 9,248us  | 25us+   |

Due to the nature of this change, performance changes from relatively constant to being proportional to the number of vehicles in the depot, simply because the number of vehicles to check changes.

<!--
Describe here shortly
* For bug fixes:
    * How is the problem solved?
* For features or gameplay changes:
    * What does this feature do?
    * How does it improve/solve the situation described under 'motivation'.
-->


## Limitations

<!--
Describe here
* Is the problem solved in all scenarios?
* Is this feature complete? Are there things that could be added in the future?
* Are there things that are intentionally left out?
* Do you know of a bug or corner case that does not work?
-->


## Checklist for review

Some things are not automated, and forgotten often. This list is a reminder for the reviewers.
* The bug fix is important enough to be backported? (label: 'backport requested')
* This PR touches english.txt or translations? Check the [guidelines](https://github.com/OpenTTD/OpenTTD/blob/master/docs/eints.md)
* This PR affects the save game format? (label 'savegame upgrade')
* This PR affects the GS/AI API? (label 'needs review: Script API')
    * ai_changelog.hpp, game_changelog.hpp need updating.
    * The compatibility wrappers (compat_*.nut) need updating.
* This PR affects the NewGRF API? (label 'needs review: NewGRF')
    * newgrf_debug_data.h may need updating.
    * [PR must be added to API tracker](https://wiki.openttd.org/en/Development/NewGRF/Specification%20Status)